ci(test): split unit vs API integration tests
