+++ src/integration-tests/headless-stream-json-format.test.ts
